Change game board to create new array of length 9
This is done largely due to the fact that later, the win conditions will revolve around where each x and o is placed. for instance if
x occupies the cells [1,2,3] that is a win condition, or if o occupies cells [0, 4, 8] that is also a win condition. instead of checking
if each cell in each row lines up perfectly for a vertical, diagonal, or horizontal, we can just check against an array of win conditions
and declare a winner or tie depending on the status of the board. It eliminates the need for nested for loops and multiple functions.