Add score tracking to the game
This adds a score property called score to the array of player objects. Then a factory function is passed to store that data. If win conditions are met
the score for the current player is updated. This score gets displayed after the round has ended. There was no message added to the score as this will later
become part of the DOM, and visual design of the page will be the indicator for who has what in scoring.