Introduce evaluate factory function in improving AI
While the initial function served an excellent staging ground for learning and understanding the logic behind minimax, the function itself when implemented into my program was fairly 'dumb'.                                                                       In order to make the AI smarter and actually play the game, I need to refactor the code to a more traditional minimax approach which
will call both the max and min methods from Math library.