Create system which allows AI to step automatically upon render
This allows for the game to respond without a click event, AND can even allow two separate AI's to duke it out. This ofcourse will require modification of the AI controller
and end states for the game to be reached. Further the computer operates to fast, and an additional 'time out' will be needed in order to facilitate animations, and give
the player the impression that this computer is really thinking about the problem.