Remove console version of board, as the board now displays in DOM
May be used later, and I am unsure if I wish to leave it as legacy code for future debugging or delete. I'm leaning towards leaving it. We can wrap a whole section of comments
at the bottom of this script file, giving instructions as to where these pieces go, should the need ever arise to return and debug code in the program