Add radio buttons for players to select the char of thier choice
In the JS we will need to modify our code to get the selected choice for each player. The default is set to
X for player one and O for player two. Of note, if both players choose the same value, one of them will need
to default to the other. This is a hardcoded decision that must be made in the next few commits. Right now the
goal is merely getting interoperability between the DOM and game.