Create variables to select the player/s choice in signs
If both are the same, player one defaults to the opposite. Also cleaned up player two ID input in HTML.
There will need to be additional arguments added to the gameController and the gameController will need to
default to either an initial symbol IE 'X' or randomly choose the first player of the round. For right now
I suggest it defaults to X.