Fix AI by passing minimizer and maximizer arguments
This eliminates the need to create a separate instance of the game controller, and can operate for both player one
and player two, as now the AI is dependent on which player was passed as max and min. There still is the issue
of the game not defending cell 8 in the array, but that can be addressed at a later date.