Add constants that get the players choice for AI's
This passes a boolean value, which is passed as arguments to the game controller. In the future an AI dificulty dropdown
menu will need to be added. This will require a random roll function within the AI controller, and might require a factory
function that has an event listener that activates when either AI radio button is clicked.