Fix 'cell8' problem
The issue was in handling of winstates and how scores were distributed. Going through
the recursive function each iteration would evaluate the player passed to the
minimax function and check for a winstate. Unfortunately this meant that the win
state would be overlooked if it was not the current player->netting in a constant
return of 0. Given, this score would get pushed to the findBestMove value and both
cells 6 and 8 gave the same score, causing the AI to just default to 6. Instead,
the new code will check both min and max arguments, and if either of those are in
a win state, the result will be returned, and an object will assign a score appropriately

Huge shoutout to Coding Train:https://thecodingtrain.com/challenges/154-tic-tac-toe-minimax
who made me realize not only the error I was making, but also how some of the code should
look. Future updates to this algorithm will include a dificulty setting, and may also include
pruning and 'subtracting' depth from the result for optimization and shorter passes.