Add a form selector for setting AI difficulty options
Gives the user an option to select the level of difficulty they wish to play against in
regards to AI. The Cheez option is unbeatable in the standard 3X3 TTT game, and uses
the minimax algorithm every turn. The next objectives are to hide this form when the
player option is selected, and show it when AI is clicked in the radio buttons. Additionally
the value for each option will need to be passed and saved into the JS player objects.
Finally, the AI function will need to be modified to determine if it should take a random
step on turn or use the minimax function, based on a threshold for difficulty. Next step
should be to hide the form until called upon by clicking on the AI radio button