Fix comment above setForm
The next step is in taking the form data and passing that as an argument to the gameController
An additional value will need to be added to the player Objects which will store a string/null property.
this string will eventually be passed to the AI function. However, we need not worry about a null value
as that will never be passed, at least not intentionally. The default that is passed when the no option is
selected should be 'easy'