Create ai precision ff skeleton
This function will take the boolean value of either AI player and return a value based on the text
content of the drop down menu. If the AI is false, our value is immediately returned as null. Else
it will switch starting with cheeze being a decimal of 1.0 to the default 'easy' with a decimal of
0.25. This will later be used to determine if the next move the AI should make is a random one or
minimax.