Fix color transitions for user name entries
The solution was to use the declaration :placeholder-shown which I surmise will only execute if the
placeholder in text is visible. The color palette is temporary, as there are other design elements
that need to be worked out first. One such element is making the radio buttons much more palatable
in style