Make radio buttons act like toggle switches
The code for this is made possible by Mike Hemberger: https://codepen.io/JiveDig . In order
for this to work several actions had to be taken.

First: each ID for inputs was changed to a unique set to match individual labels. This
allows each label to act as the button themselves.

Second: In order to accomidate this revision, all javascript query
selectors handling user names and AI selections were amended to select
the name of each input, not the ID. IE
document.querySelectorAll('[name="playerOneAI"]')