Add change to image source up Human/AI selection
Instead of bulking the code down with more images, and setting their display settings
post selection(an action that would require a lot of modification to the JS code
both with the event handler and passing of images to the gameboard) I instead opted
to simply change the source on the image during the event handling of setForm function.
This allows the switching between AI and Player images, without having to modify any
classes, ID's, or stacking conditionals in the script code.