Introduce Next Round button
This button will be handled via event, and will display at rounds
end: be it a tie or win state. I've also taken the liberty to
make all buttons display a pointer when hovered over. This
reduces the amount of redundant code in the CSS file.

The event handler still must be written into the JS code, and
the endRound() and endGame() states mush be returned and handled
by the display controller, not the game controller. For clarity,
these functions will remain and be executed via the gameController
when called, but will be handled by the display controller given
the games state(tied round, tied game...etc).