Fix issue with endless looping of AI and fix scoreboard doubling
The added variable and factory functions allow for a flag to activate when the game reaches a
tie or winstate. This flag is part of the AI conditional inside of render. the AI is halted from
executing, and the recursive call to render never happens. Until the click event for next round
is executed. There was a need to separate this flag from check, which could halt the AI call, but
would add to the score, given the flow of the code.

This commit also disposes of the winRound function, as console logging is becoming more and more
replaced with user interface. Given the compact nature of the function, I decided check could
handle winRounds responsibilities which are virtually one in the same.