Add fade in and fade out animations for AI selection form
Because setting the style to display:none immediately hides the element, it was necessary
to create a timeout function that waited to change the style. Upon execution an new class
is added. The asynce function executes counting up the 700ms needed to execute the animation
where upon the style is changed, and class removed.