Shorten animation sequence to prevent AI bug
The bug is simply clicking human then AI before the event has had time to fully execute.
The solution is to shorten the time it takes for the fadeout animation to play.
The timeout function was reduced to 0.2s. This time frame, is still beatable, but
requires the user to be deliberately trying to beat it.