Animate character inputs
It was necessary to return a value for the AI. This value is the index choice of the of
the ai function. The indices of both player and ai are passed to the render function.
In the render function I created a span and appended each to a cell. If the cell has
been 'clicked' by either party, a conditional appends a temporary class called
animate. This triggers the animation in our animation css file.